Isolate maze evolve reproducibility test from parallel evolveRL runs.
Concurrent evolveMazeController tests in the same file polluted NEAT-AI
global caches, so back-to-back runs with the same seed diverged in CI.
